pyproject.toml
[tool.poetry]
name = "aworda-lbot"
version = "0.0.2"
description = "另一个世界里，我希望你过得比这个世界的我好，霖念。"
authors = ["Little-LinNian <2544704967@qq.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.9"
graia-ariadne = "^0.8.2"
beanie = "^1.11.0"
graia-saya = "^0.0.16"
pillow = "^9.1.1"
ujson = "^5.3.0"
arclet-alconna-graia = "^0.4.5"
aiofiles = "^0.8.0"
jieba = "^0.42.1"
click = "^8.1.3"
mofish = "^1.0.4"
psutil = "^5.9.1"
prompt-toolkit = "^3.0.30"
arclet-alconna = "^1.1.2"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "txcloud"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"



